+++ src/MQTTAnalyzer/views/host/form/auth/AuthenticationTypePicker.swift
