+++ src/MQTTAnalyzer/views/host/form/auth/CertificateAuthenticationView.swift
